% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_night_records.R
\name{get_night_records}
\alias{get_night_records}
\title{get_night_records}
\usage{
get_night_records(data, start, end)
}
\arguments{
\item{data}{tsibble to select for day records}

\item{start}{time in "HH:MM:SS" format for day start}

\item{end}{time in "HH:MM:SS" format for day end}
}
\value{
night sliced tsibble containing night records as well as day and week of study from first data entry
}
\description{
get_night_records
}
\examples{
columns <- c("datetime", "value")
row1 <- c("2021-03-21 12:15:05", "bottom")
row2 <- c("2021-03-21 12:15:10", "bottom")
row3 <- c("2021-03-21 12:15:15", "bottom")
row4 <- c("2021-03-21 12:15:20", "bottom")
row5 <- c("2021-03-21 12:15:25", "bottom")
row6 <- c("2021-03-21 12:15:30", "bottom")
data2 <- data.frame(rbind(row1,row2,row3,row4,row5,row6),row.names=c())
colnames(data2) <- columns
get_night_records(data2, "04:00:00", "12:25:00")
}
