% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_budget_prop.R
\name{get_time_budget_prop}
\alias{get_time_budget_prop}
\title{calculates the proportion of time spent in each zone based on the whole time series passed in
from Analyis.R generated by Keni}
\usage{
get_time_budget_prop(data, value, options, n_opts)
}
\arguments{
\item{data}{interval table to generate time budget from}

\item{value}{name of column that contains all options (subzone)}

\item{options}{vector of options that value contains (bottom, middle, top)}

\item{n_opts}{number of options in vector (3)}
}
\value{
dataframe containing start, end and value as columns
}
\description{
calculates the proportion of time spent in each zone based on the whole time series passed in
from Analyis.R generated by Keni
}
\examples{
columns <- c("t1","t2", "from_zone", "to_zone")
rows <- c("1616328905", "1616368409", NA, "bottom")
data <- data.frame(rbind(rows))
colnames(data) <- columns
get_time_budget_prop(data)

}
