% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_duplicate_records.R
\name{identify_duplicate_records}
\alias{identify_duplicate_records}
\title{perform sliding window analysis to remove transitions into current zone}
\usage{
identify_duplicate_records(records)
}
\arguments{
\item{records}{tibble with colnamed subzone}
}
\value{
de-duplicated transtion record table
}
\description{
perform sliding window analysis to remove transitions into current zone
}
\examples{
require(dplyr)
subzone <- c("Middle","Middle","Middle","Top")
accessdate <- c(1620412230,1620411973,1620411937,1620411910)
LegBand <- c(6992,6992,6992,6992)
record_table <- data.frame(cbind(subzone,accessdate,LegBand))
colnames(record_table) <- c('subzone','accessdate','LegBand')
record_table <- dplyr::tibble(record_table)

de_duplicated_record_table <- identify_duplicate_records(record_table)
}
