% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ymd_hms.R
\name{my_ymd_hms}
\alias{my_ymd_hms}
\title{Midnight Aware YMD-HMS formatting}
\usage{
my_ymd_hms(date)
}
\arguments{
\item{date}{year-month-date T hour:minute:second date}
}
\value{
a numerically formatted datetime
}
\description{
Midnight Aware YMD-HMS formatting
}
\examples{
date_1 <- "2021-03-10 T04:00:00"
date_2 <- "2021-03-15 T00:00:00"
my_ymd_hms(date_1)
my_ymd_hms(date_2)
}
