% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_time_to_intervals.R
\name{nested_time_to_intervals}
\alias{nested_time_to_intervals}
\title{nested_time_to_intervals}
\usage{
nested_time_to_intervals(data)
}
\arguments{
\item{data}{dataframe containing records with the column dos (day of study)}
}
\value{
daily_time_budget nested interval tables, where records are nested on day of study
}
\description{
nested_time_to_intervals
}
\examples{
columns <- c("datetime","value","day","date","dos","wos")
row1 <- c("2021-03-21 12:15:05", "bottom", TRUE, 2021-03-21, 1, 1)
row2 <- c("2021-03-21 12:15:10", "bottom", TRUE, 2021-03-21, 1, 1)
row3 <- c("2021-03-21 12:15:15", "bottom", TRUE, 2021-03-21, 1, 1)
row4 <- c("2021-03-21 12:15:20", "bottom", TRUE, 2021-03-21, 1, 1)
row5 <- c("2021-03-21 12:15:25", "bottom", TRUE, 2021-03-21, 1, 1)
row6 <- c("2021-03-21 12:15:30", "bottom", TRUE, 2021-03-21, 1, 1)
data2 <- data.frame(rbind(row1,row2,row3,row4,row5,row6),row.names=c())
colnames(data2) <- columns
nested_time_to_intervals(data2)

}
