% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/night_zone_from_tb.R
\name{night_zone_from_TB}
\alias{night_zone_from_TB}
\title{night zone from time budget}
\usage{
night_zone_from_TB(data)
}
\arguments{
\item{data}{a night time budget that you want to know the zone that most time was spent in}
}
\value{
a two item vector of zones and propotional time spent in it
}
\description{
night zone from time budget
}
\examples{
columns <- c("t1","t2", "from_zone", "to_zone")
rows <- c("1616328905", "1616368409", NA, "bottom")
data <- data.frame(rbind(rows))
colnames(data) <- columns
data2 <- get_time_budget_prop(data)
night_zone_from_TB(data2)

}
