% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_tsibble.R
\name{slice_tsibble}
\alias{slice_tsibble}
\title{select Tsibble between timepoints given (used to get contiguous time series of records)}
\usage{
slice_tsibble(data, start, stop)
}
\arguments{
\item{data}{Tsibble with data between start and stop}

\item{start}{A lubridate::ymd() formatted datetime}

\item{stop}{A lubridate::ymd() formatted datetime}
}
\value{
a sliced Tsibble for values between start and stop
}
\description{
select Tsibble between timepoints given (used to get contiguous time series of records)
}
\examples{
subzone <- c("Middle","Middle","Middle","Top")
datetime <- c("2021-03-21 T23:13:00","2021-03-21 T23:12:00", "2021-03-21 T23:11:00","2021-03-21 T23:10:00")
LegBand <- c(6992,6992,6992,6992)
record_table <- data.frame(cbind(subzone,datetime,LegBand))
colnames(record_table) <- c('subzone','datetime','LegBand')
record_table <- dplyr::tibble(record_table)
end <- "2021-05-07 T18:30:30"
begin <- "2021-05-07 T18:24:10"

slice_tsibble(record_table, begin, end)
}
