% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_tsibble.R
\name{sliceTsibble}
\alias{sliceTsibble}
\title{select Tsibble between timepoints given (used to get contiguous time series of records)}
\usage{
sliceTsibble(data, start, stop)
}
\arguments{
\item{data}{Tsibble with data between start and stop}

\item{start}{A lubridate::ymd() formatted datetime}

\item{stop}{}
}
\value{
a sliced Tsibble for values between start and stop
}
\description{
select Tsibble between timepoints given (used to get contiguous time series of records)
}
