% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_intervals.R
\name{time_to_intervals}
\alias{time_to_intervals}
\title{Function to turn time table into interval table, handles case of 0,1,2,...n transitions}
\usage{
time_to_intervals(data)
}
\arguments{
\item{data}{}
}
\description{
Function to turn time table into interval table, handles case of 0,1,2,...n transitions
}
