% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_budget_prop.R
\name{get_time_budget_prop}
\alias{get_time_budget_prop}
\title{calculates the proportion of time spent in each zone based on the whole time series passed in
from Analyis.R generated by Keni}
\usage{
get_time_budget_prop(data, value, options, n_opts)
}
\arguments{
\item{data}{}
}
\value{
dataframe containing start, end and value as columns
}
\description{
calculates the proportion of time spent in each zone based on the whole time series passed in
from Analyis.R generated by Keni
}
