% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_start.R
\name{nice_start}
\alias{nice_start}
\title{Function to round the start so that records start nicely}
\usage{
nice_start(dataframe, units, interval_min, datetime)
}
\arguments{
\item{interval_min}{}
}
\description{
Function to round the start so that records start nicely
}
