% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_tsibble.R
\name{slice_tsibble}
\alias{slice_tsibble}
\title{select Tsibble between timepoints given (used to get contiguous time series of records)}
\usage{
slice_tsibble(data, start, stop, datetime = "datetime")
}
\arguments{
\item{data}{Tsibble with data between start and stop}

\item{start}{A lubridate::ymd() formatted datetime}

\item{stop}{A lubridate::ymd() formatted datetime}

\item{datetime}{String that is the columnname where the ymd_hms formatted datetime is}
}
\value{
a sliced Tsibble for values between start and stop
}
\description{
select Tsibble between timepoints given (used to get contiguous time series of records)
}
\examples{
subzone <- c("Middle","Middle","Middle","Top")
datetime <- c("2021-03-21 T23:13:00","2021-03-21 T23:12:00", "2021-03-21 T23:11:00","2021-03-21 T23:10:00")
accessdate <- c("2021-03-21 T23:13:00","2021-03-21 T23:12:00", "2021-03-21 T23:11:00","2021-03-21 T23:10:00")
LegBand <- c(6992,6992,6992,6992)
record_table <- data.frame(cbind(subzone,datetime,accessdate,LegBand))
colnames(record_table) <- c('subzone','datetime','accessdate','LegBand')
record_table <- dplyr::tibble(record_table)
begin <- "2021-03-21 T23:10:00"
end <- "2021-03-21 T23:12:00"

slice_tsibble(record_table, start=begin, stop=end)
}
